orderbook-processor/requirements.txt
kafka-python==2.0.2
clickhouse-connect==0.7.19
PyYAML==6.0.2
